+++ cypress/e2e/sessions.cy.js
@@ -30,7 +30,7 @@ const dummyQueue = [
 describe("Session tests ", () => {
     it("Checks if session start, extension and ending works with a dummy queue", () => {
         hp.haltAndClearStorage(() => {
-        // initialize countly
+            // initialize countly
             initMain();
             // begin session
             Countly.begin_session();
@@ -50,7 +50,7 @@ describe("Session tests ", () => {
     });
     it("Checks if session start, extension and ending works", () => {
         hp.haltAndClearStorage(() => {
-        // initialize countly
+            // initialize countly
             initMain();
             // begin session
             Countly.begin_session();
@@ -97,9 +97,7 @@ describe("Browser session tests, auto", () => {
 describe("Browser session tests, manual 1", () => {
     it("Single sessions test with manual sessions", () => {
         cy.visit("./cypress/fixtures/session_test_manual_1.html?use_session_cookie=true");
-        cy.contains("Start").click().wait(waitTime);
-        cy.contains("Event").click().wait(300);
-        cy.contains("End").click().wait(300);
+        cy.wait(waitTime + 1000);
         cy.visit("./cypress/fixtures/base.html");
         cy.fetch_local_request_queue(app_key).then((rq) => {
             cy.log(rq);
@@ -109,10 +107,10 @@ describe("Browser session tests, manual 1", () => {
             cy.check_session(rq[0], undefined, undefined, app_key);
             // third object of the queue should be about session extension, also input the expected duration
             cy.check_session(rq[2], 5, undefined, app_key);
-            // fourth object of the queue should be about event sent
-            cy.check_event(JSON.parse(rq[3].events)[0], eventObj, undefined, false);
             // fifth object of the queue should be about session extension, also input the expected duration
-            cy.check_session(rq[4], 1, undefined, app_key);
+            cy.check_session(rq[3], 1, undefined, app_key);
+            // fourth object of the queue should be about event sent
+            cy.check_event(JSON.parse(rq[4].events)[0], eventObj, undefined, false);
         });
     });
 });
@@ -160,12 +158,7 @@ describe("Browser session tests, auto, no cookie", () => {
 describe("Browser session tests, manual 1, no cookie", () => {
     it("Single bounce test with manual sessions with no cookies", () => {
         cy.visit("./cypress/fixtures/session_test_manual_1.html");
-        cy.contains("Start").click();
-        cy.wait(waitTime);
-        cy.contains("Event").click();
-        cy.wait(300);
-        cy.contains("End").click();
-        cy.wait(300);
+        cy.wait(waitTime + 1000);
         cy.visit("./cypress/fixtures/base.html");
         cy.fetch_local_request_queue(app_key).then((rq) => {
             cy.log(rq);
@@ -175,10 +168,10 @@ describe("Browser session tests, manual 1, no cookie", () => {
             cy.check_session(rq[0], undefined, undefined, app_key);
             // third object of the queue should be about session extension, also input the expected duration
             cy.check_session(rq[2], 5, undefined, app_key);
-            // fourth object of the queue should be about event sent
-            cy.check_event(JSON.parse(rq[3].events)[0], eventObj, undefined, false);
             // fifth object of the queue should be about session extension, also input the expected duration
-            cy.check_session(rq[4], 1, true, app_key);
+            cy.check_session(rq[3], 1, true, app_key);
+            // fourth object of the queue should be about event sent
+            cy.check_event(JSON.parse(rq[4].events)[0], eventObj, undefined, false);
         });
     });
 });
@@ -205,7 +198,7 @@ describe("Browser session tests, manual 2, no cookie", () => {
 describe("Check request related functions", () => {
     it("Check if prepareRequest forms a proper request object", () => {
         hp.haltAndClearStorage(() => {
-        // initialize countly
+            // initialize countly
             initMain();
             let reqObject = {};
             Countly._internals.prepareRequest(reqObject);
@@ -215,7 +208,7 @@ describe("Check request related functions", () => {
     });
     it("Check if prepareRequest forms a proper request object from a bad one ", () => {
         hp.haltAndClearStorage(() => {
-        // initialize countly
+            // initialize countly
             initMain();
             let reqObject = { app_key: null, device_id: null };
             Countly._internals.prepareRequest(reqObject);
@@ -225,7 +218,7 @@ describe("Check request related functions", () => {
     });
     it("Check if prepareRequest forms a proper request object and not erase an extra value ", () => {
         hp.haltAndClearStorage(() => {
-        // initialize countly
+            // initialize countly
             initMain();
             let reqObject = { extraKey: "value" };
             Countly._internals.prepareRequest(reqObject);