+++ lib/countly.d.ts
@@ -484,6 +484,22 @@ declare module "countly-sdk-web" {
     }
 
     /**
+     * Content interface with convenience methods for content zones:
+     * - enterContentZone() - enters a content zone
+     * - exitContentZone() - exits a content zone
+     */
+    const content: Content;
+    interface Content {
+      /**
+       * Enters content zone and checks and displays available content regularly
+       */
+      enterContentZone(): void;
+      /**
+       * Exits content zone
+       */
+      exitContentZone(): void;
+    }
+    /**
      * This function retrieves all associated widget information (IDs, type, name etc in an array/list of objects) of your app
      * @param {Function} callback - Callback function with two parameters, 1st for returned list, 2nd for error
      * @returns {void}